+++ packages/core/src/transaction/TransactionBody.ts
@@ -1,51 +1,4 @@
-/**
- * Simple type for transaction clause.
- */
-interface TransactionClause {
-    /**
-     * Destination address where:
-     * * transfer token to or
-     * * invoke contract method on.
-     *
-     * @note Set null destination to deploy a contract.
-     */
-    to: string | null;
-
-    /**
-     * Amount of token to transfer to the destination
-     */
-    value: string | number;
-
-    /**
-     * Input data for contract method invocation or deployment
-     */
-    data: string;
-
-    /**
-     * Optional comment for the clause, helpful for displaying what the clause is doing.
-     */
-    comment?: string;
-
-    /**
-     * Optional ABI for the contract method invocation.
-     */
-    abi?: string;
-}
-
-/**
- * Options for creating a clause.
- */
-interface ClauseOptions {
-    /**
-     * Optional comment for the clause, helpful for displaying what the clause is doing.
-     */
-    comment?: string;
-
-    /**
-     * Optional ABI for the contract method invocation.
-     */
-    includeABI?: boolean;
-}
+import { type TransactionClause } from './TransactionClause';
 
 /**
  * Type for transaction body.
@@ -135,4 +88,4 @@ interface TransactionBody {
     };
 }
 
-export type { TransactionBody, TransactionClause, ClauseOptions };
+export type { TransactionBody };
