+++ packages/core/src/transaction/index.ts
@@ -1,6 +1,8 @@
 export * from './Clause';
+export type * from './ClauseOptions';
 export type * from './DeployParams';
+export type * from './TransactionBody';
+export type * from './TransactionClause';
 export * from './transaction';
 export * from './transaction-handler';
-export type * from './types.d';
 export * from '../utils/const/network';
